BigProject/clinic_manager/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cloudinary==1.39.1
DateTime==5.5
Django==5.0.4
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
mysqlclient==2.2.4
packaging==24.0
pytz==2024.1
PyYAML==6.0.1
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
zope.interface==6.2
